Revert "Move experimental.kernel_iface into top-level module."
This reverts commit 790e4a05fbd8f04cf456f2bfbd140afce21ac287.
